Added Prototype 1
Added the prototype one so that one a special ability is picked up it will enable the 2nd ability slot. To add more abilities, one should create a new scriptable object (inside Drops) with a new name and in the (Spells) one can add another object there with a different icon / name / properties.